backend/requirements.txt
# Backend Requirements for Resume AI Platform

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0

# NLP & ML
spacy==3.7.2
torch==2.0.1
sentence-transformers==2.2.2
transformers==4.35.2
faiss-cpu==1.7.4

# File Processing
PyPDF2==3.0.1
python-docx==0.8.11
pypdf==3.17.1

# Caching & Async
redis==5.0.1
celery==5.3.4

# File Storage
minio==7.2.0

# Utilities
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
regex==2023.12.25

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging & Monitoring
python-json-logger==2.0.7

# API Documentation
typing-extensions==4.8.0
